remnote-flashcard-generator/requirements.txt
pyyaml>=6.0
jsonschema>=4.0
openai>=1.0
python-dotenv>=1.0
click>=8.0
rich>=13.0
pydantic>=2.0
anthropic>=0.25.0
tiktoken>=0.5.0

# Optional testing dependencies
pytest>=7.0
pytest-mock>=3.10
pytest-cov>=4.0
